services/crawler/requirements.txt
praw
pyyaml
boto3